requirements.txt
asgiref==3.5.0
Django==4.0.4
image==1.5.33
numpy==1.22.3
Pillow==9.1.0
six==1.16.0
sqlparse==0.4.2